refactor: split file transfer failure boundaries
