+++ backend/internal/api/file_transfer_runner.go
@@ -5,15 +5,12 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"time"
 
 	"github.com/aipermission/aipermission/backend/internal/connectorapi"
 	"github.com/aipermission/aipermission/backend/internal/filetransfer"
 )
-
-var errFileTransferTimedOut = errors.New("file transfer timed out")
 
 func (s fileTransferHandlers) runUpload(runtime *databaseRuntime, transferID int64, overwrite bool) {
 	ctx, cancel := context.WithTimeout(context.Background(), fileTransferTimeout)
@@ -330,175 +327,4 @@ func (s fileTransferHandlers) transferProgress(runtime *databaseRuntime, transfe
 			}
 		}
 	}
-}
-
-func (s fileTransferHandlers) checkUploadOverwrite(w http.ResponseWriter, r *http.Request, runtime *databaseRuntime, runtimeID int64, remotePath string, overwrite bool, tempPath string) bool {
-	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
-	defer cancel()
-	adapter, err := s.fileTransferAdapter(ctx, runtime, runtimeID)
-	if err != nil {
-		_ = os.Remove(tempPath)
-		handleConnectorTargetRuntimeError(w, err)
-		return false
-	}
-	status, err := adapter.StatRemotePath(ctx, s.Server, runtime, runtimeID, remotePath)
-	if err != nil {
-		_ = os.Remove(tempPath)
-		if writeConnectorError(w, adapter, err) {
-			return false
-		}
-		writeError(w, http.StatusBadGateway, connectorErrorMessage(adapter, "remote path check failed", err))
-		return false
-	}
-	if !status.Exists {
-		return true
-	}
-	if status.Type != "file" {
-		_ = os.Remove(tempPath)
-		writeJSON(w, http.StatusConflict, remoteFileExistsResponse{
-			Error:      "remote path already exists and is not a regular file",
-			Code:       "remote_path_exists",
-			RemotePath: remotePath,
-			Type:       status.Type,
-			Size:       status.Size,
-		})
-		return false
-	}
-	if !overwrite {
-		_ = os.Remove(tempPath)
-		writeJSON(w, http.StatusConflict, remoteFileExistsResponse{
-			Error:      "remote file already exists",
-			Code:       "remote_file_exists",
-			RemotePath: remotePath,
-			Type:       status.Type,
-			Size:       status.Size,
-		})
-		return false
-	}
-	return true
-}
-
-func (s fileTransferHandlers) checkUploadBatchOverwrite(w http.ResponseWriter, r *http.Request, runtime *databaseRuntime, runtimeID int64, requests []filetransfer.CreateRequest, overwrite bool, tempPaths []string) ([]remoteFileConflict, bool) {
-	ctx, cancel := context.WithTimeout(r.Context(), 60*time.Second)
-	defer cancel()
-	adapter, err := s.fileTransferAdapter(ctx, runtime, runtimeID)
-	if err != nil {
-		cleanupTempPaths(tempPaths)
-		handleConnectorTargetRuntimeError(w, err)
-		return nil, false
-	}
-	var conflicts []remoteFileConflict
-	for _, item := range requests {
-		status, err := adapter.StatRemotePath(ctx, s.Server, runtime, runtimeID, item.RemotePath)
-		if err != nil {
-			cleanupTempPaths(tempPaths)
-			if writeConnectorError(w, adapter, err) {
-				return nil, false
-			}
-			writeError(w, http.StatusBadGateway, connectorErrorMessage(adapter, "remote path check failed", err))
-			return nil, false
-		}
-		if !status.Exists {
-			continue
-		}
-		if status.Type != "file" {
-			cleanupTempPaths(tempPaths)
-			writeJSON(w, http.StatusConflict, remoteFileConflictsResponse{
-				Error: "one or more remote paths already exist and are not regular files",
-				Code:  "remote_paths_exist",
-				Conflicts: []remoteFileConflict{{
-					RemotePath: item.RemotePath,
-					Type:       status.Type,
-					Size:       status.Size,
-				}},
-			})
-			return nil, false
-		}
-		if !overwrite {
-			conflicts = append(conflicts, remoteFileConflict{RemotePath: item.RemotePath, Type: status.Type, Size: status.Size})
-		}
-	}
-	if len(conflicts) > 0 {
-		cleanupTempPaths(tempPaths)
-		return conflicts, false
-	}
-	return nil, true
-}
-
-func (s fileTransferHandlers) failFileTransfer(runtime *databaseRuntime, transferID int64, err error, failureKind string) {
-	message := fileTransferFailureMessage(err)
-	changed, writeErr := runtime.fileTransfers.FailWithKind(context.Background(), transferID, message, failureKind)
-	if writeErr != nil {
-		log.Printf("fail file transfer failed transfer=%d error=%v", transferID, writeErr)
-	}
-	if !changed {
-		return
-	}
-}
-
-func (s fileTransferHandlers) finishFileTransferError(runtime *databaseRuntime, transferID int64, ctx context.Context, err error) {
-	switch classifyFileTransferInterruption(ctx, err) {
-	case fileTransferTimedOut:
-		s.failFileTransfer(runtime, transferID, errFileTransferTimedOut, filetransfer.FailureKindTimeout)
-	case fileTransferCanceledByUser:
-		s.cancelFileTransferRecord(runtime, transferID, "canceled by local user")
-	default:
-		s.failFileTransfer(runtime, transferID, err, filetransfer.FailureKindUnknown)
-	}
-}
-
-func (s fileTransferHandlers) finishFileTransferBatchError(runtime *databaseRuntime, batchID int64, ctx context.Context, err error) {
-	switch classifyFileTransferInterruption(ctx, err) {
-	case fileTransferTimedOut:
-		if _, writeErr := runtime.fileTransfers.FailBatchWithKind(context.Background(), batchID, "file transfer batch timed out", filetransfer.FailureKindTimeout); writeErr != nil {
-			log.Printf("mark file transfer batch timed out failed batch=%d error=%v", batchID, writeErr)
-		}
-	case fileTransferCanceledByUser:
-		if _, writeErr := runtime.fileTransfers.CancelBatch(context.Background(), batchID, "canceled by local user"); writeErr != nil {
-			log.Printf("cancel file transfer batch failed batch=%d error=%v", batchID, writeErr)
-		}
-	default:
-		if _, writeErr := runtime.fileTransfers.FailBatchWithKind(context.Background(), batchID, fileTransferFailureMessage(err), filetransfer.FailureKindUnknown); writeErr != nil {
-			log.Printf("fail file transfer batch failed batch=%d error=%v", batchID, writeErr)
-		}
-	}
-}
-
-type fileTransferInterruption int
-
-const (
-	fileTransferNotInterrupted fileTransferInterruption = iota
-	fileTransferCanceledByUser
-	fileTransferTimedOut
-)
-
-func classifyFileTransferInterruption(ctx context.Context, err error) fileTransferInterruption {
-	if errors.Is(ctx.Err(), context.DeadlineExceeded) || errors.Is(err, context.DeadlineExceeded) {
-		return fileTransferTimedOut
-	}
-	if errors.Is(ctx.Err(), context.Canceled) {
-		return fileTransferCanceledByUser
-	}
-	return fileTransferNotInterrupted
-}
-
-func (s fileTransferHandlers) cancelFileTransferRecord(runtime *databaseRuntime, transferID int64, message string) {
-	changed, err := runtime.fileTransfers.Cancel(context.Background(), transferID, message)
-	if err != nil {
-		log.Printf("cancel file transfer failed transfer=%d error=%v", transferID, err)
-		return
-	}
-	if !changed {
-		return
-	}
-}
-
-func fileTransferFailureMessage(err error) string {
-	if err == nil {
-		return ""
-	}
-	if errors.Is(err, errFileTransferTimedOut) {
-		return errFileTransferTimedOut.Error()
-	}
-	return fmt.Sprintf("file transfer failed: %v", err)
 }
